+++ server/mux.go
@@ -143,8 +143,15 @@ func NewMux(opts MuxOpts, service Service) http.Handler {
 	router.PUT("/chronograf/v1/layouts/:id", service.UpdateLayout)
 	router.DELETE("/chronograf/v1/layouts/:id", service.RemoveLayout)
 
-	// Users
+	// Users associated with Chronograf
 	router.GET("/chronograf/v1/me", service.Me)
+
+	// router.GET("/chronograf/v1/users", service.Users)
+	router.POST("/chronograf/v1/users", service.NewUser)
+
+	router.GET("/chronograf/v1/users/:id", service.UserID)
+	router.DELETE("/chronograf/v1/users/:id", service.RemoveUser)
+	router.PATCH("/chronograf/v1/users/:id", service.UpdateUser)
 
 	// Dashboards
 	router.GET("/chronograf/v1/dashboards", service.Dashboards)
