+++ server/users_test.go
@@ -1,8 +1,9 @@
-package server_test
+package server
 
 import (
 	"bytes"
 	"context"
+	"encoding/json"
 	"fmt"
 	"io/ioutil"
 	"net/http"
@@ -10,18 +11,16 @@ import (
 	"testing"
 
 	"github.com/bouk/httprouter"
+	"github.com/google/go-cmp/cmp"
 	"github.com/influxdata/chronograf"
 	"github.com/influxdata/chronograf/log"
 	"github.com/influxdata/chronograf/mocks"
-	"github.com/influxdata/chronograf/server"
 )
 
-func TestService_NewSourceUser(t *testing.T) {
+func TestService_UserID(t *testing.T) {
 	type fields struct {
-		SourcesStore chronograf.SourcesStore
-		TimeSeries   server.TimeSeriesClient
-		Logger       chronograf.Logger
-		UseAuth      bool
+		UsersStore chronograf.UsersStore
+		Logger     chronograf.Logger
 	}
 	type args struct {
 		w *httptest.ResponseRecorder
@@ -31,900 +30,356 @@ func TestService_NewSourceUser(t *testing.T) {
 		name            string
 		fields          fields
 		args            args
-		ID              string
+		id              string
 		wantStatus      int
 		wantContentType string
 		wantBody        string
 	}{
 		{
-			name: "New user for data source",
+			name: "Get Single Chronograf User",
+			args: args{
+				w: httptest.NewRecorder(),
+				r: httptest.NewRequest(
+					"GET",
+					"http://any.url", // can be any valid URL as we are bypassing mux
+					nil,
+				),
+			},
+			fields: fields{
+				Logger: log.New(log.DebugLevel),
+				UsersStore: &mocks.UsersStore{
+					GetF: func(ctx context.Context, ID string) (*chronograf.User, error) {
+						switch ID {
+						case "1337":
+							return &chronograf.User{
+								ID:       1337,
+								Name:     "billysteve",
+								Provider: "Google",
+								Scheme:   "OAuth2",
+							}, nil
+						default:
+							return nil, fmt.Errorf("User with ID %v not found", ID)
+						}
+					},
+				},
+			},
+			id:              "1337",
+			wantStatus:      http.StatusOK,
+			wantContentType: "application/json",
+			wantBody:        `{"id":1337,"name":"billysteve","provider":"Google","scheme":"OAuth2","links":{"self":"/chronograf/v1/users/1337"}}`,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			s := &Service{
+				UsersStore: tt.fields.UsersStore,
+				Logger:     tt.fields.Logger,
+			}
+
+			tt.args.r = tt.args.r.WithContext(httprouter.WithParams(
+				context.Background(),
+				httprouter.Params{
+					{
+						Key:   "id",
+						Value: tt.id,
+					},
+				}))
+
+			s.UserID(tt.args.w, tt.args.r)
+
+			resp := tt.args.w.Result()
+			content := resp.Header.Get("Content-Type")
+			body, _ := ioutil.ReadAll(resp.Body)
+
+			if resp.StatusCode != tt.wantStatus {
+				t.Errorf("%q. UserID() = %v, want %v", tt.name, resp.StatusCode, tt.wantStatus)
+			}
+			if tt.wantContentType != "" && content != tt.wantContentType {
+				t.Errorf("%q. UserID() = %v, want %v", tt.name, content, tt.wantContentType)
+			}
+			if eq, _ := jsonEqual(string(body), tt.wantBody); tt.wantBody != "" && !eq {
+				t.Errorf("%q. UserID() = \n***%v***\n,\nwant\n***%v***", tt.name, string(body), tt.wantBody)
+			}
+		})
+	}
+}
+
+func TestService_NewUser(t *testing.T) {
+	type fields struct {
+		UsersStore chronograf.UsersStore
+		Logger     chronograf.Logger
+	}
+	type args struct {
+		w    *httptest.ResponseRecorder
+		r    *http.Request
+		user chronograf.User
+	}
+	tests := []struct {
+		name            string
+		fields          fields
+		args            args
+		wantStatus      int
+		wantContentType string
+		wantBody        string
+	}{
+		{
+			name: "Create a new Chronograf User",
 			args: args{
 				w: httptest.NewRecorder(),
 				r: httptest.NewRequest(
 					"POST",
-					"http://server.local/chronograf/v1/sources/1",
-					ioutil.NopCloser(
-						bytes.NewReader([]byte(`{"name": "marty", "password": "the_lake"}`)))),
+					"http://any.url",
+					nil,
+				),
+				user: chronograf.User{
+					Name:     "bob",
+					Provider: "GitHub",
+					Scheme:   "OAuth2",
+				},
 			},
 			fields: fields{
-				UseAuth: true,
-				Logger:  log.New(log.DebugLevel),
-				SourcesStore: &mocks.SourcesStore{
-					GetF: func(ctx context.Context, ID int) (chronograf.Source, error) {
-						return chronograf.Source{
-							ID:       1,
-							Name:     "muh source",
-							Username: "name",
-							Password: "hunter2",
-							URL:      "http://localhost:8086",
+				Logger: log.New(log.DebugLevel),
+				UsersStore: &mocks.UsersStore{
+					AddF: func(ctx context.Context, user *chronograf.User) (*chronograf.User, error) {
+						return &chronograf.User{
+							ID:       1338,
+							Name:     "bob",
+							Provider: "GitHub",
+							Scheme:   "OAuth2",
 						}, nil
 					},
 				},
-				TimeSeries: &mocks.TimeSeries{
-					ConnectF: func(ctx context.Context, src *chronograf.Source) error {
-						return nil
-					},
-					UsersF: func(ctx context.Context) chronograf.UsersStore {
-						return &mocks.UsersStore{
-							AddF: func(ctx context.Context, u *chronograf.User) (*chronograf.User, error) {
-								return u, nil
-							},
-						}
-					},
-					RolesF: func(ctx context.Context) (chronograf.RolesStore, error) {
-						return nil, fmt.Errorf("no roles")
-					},
-				},
 			},
-			ID:              "1",
 			wantStatus:      http.StatusCreated,
 			wantContentType: "application/json",
-			wantBody: `{"links":{"self":"/chronograf/v1/sources/1/users/marty"},"name":"marty","permissions":[]}
-`,
-		},
-		{
-			name: "New user for data source with roles",
-			args: args{
-				w: httptest.NewRecorder(),
-				r: httptest.NewRequest(
-					"POST",
-					"http://server.local/chronograf/v1/sources/1",
-					ioutil.NopCloser(
-						bytes.NewReader([]byte(`{"name": "marty", "password": "the_lake"}`)))),
-			},
-			fields: fields{
-				UseAuth: true,
-				Logger:  log.New(log.DebugLevel),
-				SourcesStore: &mocks.SourcesStore{
-					GetF: func(ctx context.Context, ID int) (chronograf.Source, error) {
-						return chronograf.Source{
-							ID:       1,
-							Name:     "muh source",
-							Username: "name",
-							Password: "hunter2",
-							URL:      "http://localhost:8086",
-						}, nil
-					},
-				},
-				TimeSeries: &mocks.TimeSeries{
-					ConnectF: func(ctx context.Context, src *chronograf.Source) error {
-						return nil
-					},
-					UsersF: func(ctx context.Context) chronograf.UsersStore {
-						return &mocks.UsersStore{
-							AddF: func(ctx context.Context, u *chronograf.User) (*chronograf.User, error) {
-								return u, nil
-							},
-						}
-					},
-					RolesF: func(ctx context.Context) (chronograf.RolesStore, error) {
-						return nil, nil
-					},
-				},
-			},
-			ID:              "1",
-			wantStatus:      http.StatusCreated,
-			wantContentType: "application/json",
-			wantBody: `{"links":{"self":"/chronograf/v1/sources/1/users/marty"},"name":"marty","permissions":[],"roles":[]}
-`,
-		},
-		{
-			name: "Error adding user",
-			args: args{
-				w: httptest.NewRecorder(),
-				r: httptest.NewRequest(
-					"POST",
-					"http://server.local/chronograf/v1/sources/1",
-					ioutil.NopCloser(
-						bytes.NewReader([]byte(`{"name": "marty", "password": "the_lake"}`)))),
-			},
-			fields: fields{
-				UseAuth: true,
-				Logger:  log.New(log.DebugLevel),
-				SourcesStore: &mocks.SourcesStore{
-					GetF: func(ctx context.Context, ID int) (chronograf.Source, error) {
-						return chronograf.Source{
-							ID:       1,
-							Name:     "muh source",
-							Username: "name",
-							Password: "hunter2",
-							URL:      "http://localhost:8086",
-						}, nil
-					},
-				},
-				TimeSeries: &mocks.TimeSeries{
-					ConnectF: func(ctx context.Context, src *chronograf.Source) error {
-						return nil
-					},
-					UsersF: func(ctx context.Context) chronograf.UsersStore {
-						return &mocks.UsersStore{
-							AddF: func(ctx context.Context, u *chronograf.User) (*chronograf.User, error) {
-								return nil, fmt.Errorf("Weight Has Nothing to Do With It")
-							},
-						}
-					},
-				},
-			},
-			ID:              "1",
-			wantStatus:      http.StatusBadRequest,
-			wantContentType: "application/json",
-			wantBody:        `{"code":400,"message":"Weight Has Nothing to Do With It"}`,
-		},
-		{
-			name: "Failure connecting to user store",
-			args: args{
-				w: httptest.NewRecorder(),
-				r: httptest.NewRequest(
-					"POST",
-					"http://server.local/chronograf/v1/sources/1",
-					ioutil.NopCloser(
-						bytes.NewReader([]byte(`{"name": "marty", "password": "the_lake"}`)))),
-			},
-			fields: fields{
-				UseAuth: true,
-				Logger:  log.New(log.DebugLevel),
-				SourcesStore: &mocks.SourcesStore{
-					GetF: func(ctx context.Context, ID int) (chronograf.Source, error) {
-						return chronograf.Source{
-							ID:       1,
-							Name:     "muh source",
-							Username: "name",
-							Password: "hunter2",
-							URL:      "http://localhost:8086",
-						}, nil
-					},
-				},
-				TimeSeries: &mocks.TimeSeries{
-					ConnectF: func(ctx context.Context, src *chronograf.Source) error {
-						return fmt.Errorf("Biff just happens to be my supervisor")
-					},
-				},
-			},
-			ID:              "1",
-			wantStatus:      http.StatusBadRequest,
-			wantContentType: "application/json",
-			wantBody:        `{"code":400,"message":"Unable to connect to source 1: Biff just happens to be my supervisor"}`,
-		},
-		{
-			name: "Failure getting source",
-			args: args{
-				w: httptest.NewRecorder(),
-				r: httptest.NewRequest(
-					"POST",
-					"http://server.local/chronograf/v1/sources/1",
-					ioutil.NopCloser(
-						bytes.NewReader([]byte(`{"name": "marty", "password": "the_lake"}`)))),
-			},
-			fields: fields{
-				UseAuth: true,
-				Logger:  log.New(log.DebugLevel),
-				SourcesStore: &mocks.SourcesStore{
-					GetF: func(ctx context.Context, ID int) (chronograf.Source, error) {
-						return chronograf.Source{}, fmt.Errorf("No McFly ever amounted to anything in the history of Hill Valley")
-					},
-				},
-			},
-			ID:              "1",
-			wantStatus:      http.StatusNotFound,
-			wantContentType: "application/json",
-			wantBody:        `{"code":404,"message":"ID 1 not found"}`,
-		},
-		{
-			name: "Bad ID",
-			args: args{
-				w: httptest.NewRecorder(),
-				r: httptest.NewRequest(
-					"POST",
-					"http://server.local/chronograf/v1/sources/1",
-					ioutil.NopCloser(
-						bytes.NewReader([]byte(`{"name": "marty", "password": "the_lake"}`)))),
-			},
-			fields: fields{
-				UseAuth: true,
-				Logger:  log.New(log.DebugLevel),
-			},
-			ID:              "BAD",
-			wantStatus:      http.StatusUnprocessableEntity,
-			wantContentType: "application/json",
-			wantBody:        `{"code":422,"message":"Error converting ID BAD"}`,
-		},
-		{
-			name: "Bad name",
-			args: args{
-				w: httptest.NewRecorder(),
-				r: httptest.NewRequest(
-					"POST",
-					"http://server.local/chronograf/v1/sources/1",
-					ioutil.NopCloser(
-						bytes.NewReader([]byte(`{"password": "the_lake"}`)))),
-			},
-			fields: fields{
-				UseAuth: true,
-				Logger:  log.New(log.DebugLevel),
-			},
-			ID:              "BAD",
-			wantStatus:      http.StatusUnprocessableEntity,
-			wantContentType: "application/json",
-			wantBody:        `{"code":422,"message":"Username required"}`,
-		},
-		{
-			name: "Bad JSON",
-			args: args{
-				w: httptest.NewRecorder(),
-				r: httptest.NewRequest(
-					"POST",
-					"http://server.local/chronograf/v1/sources/1",
-					ioutil.NopCloser(
-						bytes.NewReader([]byte(`{password}`)))),
-			},
-			fields: fields{
-				UseAuth: true,
-				Logger:  log.New(log.DebugLevel),
-			},
-			ID:              "BAD",
-			wantStatus:      http.StatusBadRequest,
-			wantContentType: "application/json",
-			wantBody:        `{"code":400,"message":"Unparsable JSON"}`,
+			wantBody:        `{"id":1338,"name":"bob","provider":"GitHub","scheme":"OAuth2","links":{"self":"/chronograf/v1/users/1338"}}`,
 		},
 	}
+
 	for _, tt := range tests {
-		tt.args.r = tt.args.r.WithContext(httprouter.WithParams(
-			context.Background(),
-			httprouter.Params{
-				{
-					Key:   "id",
-					Value: tt.ID,
-				},
-			}))
+		t.Run(tt.name, func(t *testing.T) {
+			s := &Service{
+				UsersStore: tt.fields.UsersStore,
+				Logger:     tt.fields.Logger,
+			}
 
-		h := &server.Service{
-			SourcesStore:     tt.fields.SourcesStore,
-			TimeSeriesClient: tt.fields.TimeSeries,
-			Logger:           tt.fields.Logger,
-			UseAuth:          tt.fields.UseAuth,
-		}
+			buf, _ := json.Marshal(tt.args.user)
+			tt.args.r.Body = ioutil.NopCloser(bytes.NewReader(buf))
 
-		h.NewSourceUser(tt.args.w, tt.args.r)
+			s.NewUser(tt.args.w, tt.args.r)
 
-		resp := tt.args.w.Result()
-		content := resp.Header.Get("Content-Type")
-		body, _ := ioutil.ReadAll(resp.Body)
+			resp := tt.args.w.Result()
+			content := resp.Header.Get("Content-Type")
+			body, _ := ioutil.ReadAll(resp.Body)
 
-		if resp.StatusCode != tt.wantStatus {
-			t.Errorf("%q. NewSourceUser() = %v, want %v", tt.name, resp.StatusCode, tt.wantStatus)
-		}
-		if tt.wantContentType != "" && content != tt.wantContentType {
-			t.Errorf("%q. NewSourceUser() = %v, want %v", tt.name, content, tt.wantContentType)
-		}
-		if tt.wantBody != "" && string(body) != tt.wantBody {
-			t.Errorf("%q. NewSourceUser() = \n***%v***\n,\nwant\n***%v***", tt.name, string(body), tt.wantBody)
-		}
+			if resp.StatusCode != tt.wantStatus {
+				t.Errorf("%q. UserID() = %v, want %v", tt.name, resp.StatusCode, tt.wantStatus)
+			}
+			if tt.wantContentType != "" && content != tt.wantContentType {
+				t.Errorf("%q. UserID() = %v, want %v", tt.name, content, tt.wantContentType)
+			}
+			if eq, _ := jsonEqual(string(body), tt.wantBody); tt.wantBody != "" && !eq {
+				t.Errorf("%q. UserID() = \n***%v***\n,\nwant\n***%v***", tt.name, string(body), tt.wantBody)
+			}
+		})
 	}
 }
 
-func TestService_SourceUsers(t *testing.T) {
+func TestService_RemoveUser(t *testing.T) {
 	type fields struct {
-		SourcesStore chronograf.SourcesStore
-		TimeSeries   server.TimeSeriesClient
-		Logger       chronograf.Logger
-		UseAuth      bool
+		UsersStore chronograf.UsersStore
+		Logger     chronograf.Logger
 	}
 	type args struct {
-		w *httptest.ResponseRecorder
-		r *http.Request
+		w    *httptest.ResponseRecorder
+		r    *http.Request
+		user chronograf.User
 	}
 	tests := []struct {
-		name            string
-		fields          fields
-		args            args
-		ID              string
-		wantStatus      int
-		wantContentType string
-		wantBody        string
+		name       string
+		fields     fields
+		args       args
+		user       *chronograf.User
+		id         string
+		wantStatus int
 	}{
 		{
-			name: "All users for data source",
+			name: "Delete a Chronograf User",
+			fields: fields{
+				Logger: log.New(log.DebugLevel),
+				UsersStore: &mocks.UsersStore{
+					GetF: func(ctx context.Context, ID string) (*chronograf.User, error) {
+						switch ID {
+						case "1339":
+							return &chronograf.User{
+								ID:       1339,
+								Name:     "helena",
+								Provider: "Heroku",
+								Scheme:   "LDAP",
+							}, nil
+						default:
+							return nil, fmt.Errorf("User with ID %v not found", ID)
+						}
+					},
+					DeleteF: func(ctx context.Context, user *chronograf.User) error {
+						return nil
+					},
+				},
+			},
 			args: args{
 				w: httptest.NewRecorder(),
 				r: httptest.NewRequest(
-					"GET",
-					"http://server.local/chronograf/v1/sources/1",
-					nil),
-			},
-			fields: fields{
-				UseAuth: true,
-				Logger:  log.New(log.DebugLevel),
-				SourcesStore: &mocks.SourcesStore{
-					GetF: func(ctx context.Context, ID int) (chronograf.Source, error) {
-						return chronograf.Source{
-							ID:       1,
-							Name:     "muh source",
-							Username: "name",
-							Password: "hunter2",
-							URL:      "http://localhost:8086",
-						}, nil
-					},
-				},
-				TimeSeries: &mocks.TimeSeries{
-					ConnectF: func(ctx context.Context, src *chronograf.Source) error {
-						return nil
-					},
-					RolesF: func(ctx context.Context) (chronograf.RolesStore, error) {
-						return nil, fmt.Errorf("no roles")
-					},
-					UsersF: func(ctx context.Context) chronograf.UsersStore {
-						return &mocks.UsersStore{
-							AllF: func(ctx context.Context) ([]chronograf.User, error) {
-								return []chronograf.User{
-									{
-										Name:   "strickland",
-										Passwd: "discipline",
-										Permissions: chronograf.Permissions{
-											{
-												Scope:   chronograf.AllScope,
-												Allowed: chronograf.Allowances{"READ"},
-											},
-										},
-									},
-								}, nil
-							},
-						}
-					},
+					"DELETE",
+					"http://any.url",
+					nil,
+				),
+				user: chronograf.User{
+					ID:       1339,
+					Name:     "helena",
+					Provider: "Heroku",
+					Scheme:   "LDAP",
 				},
 			},
-			ID:              "1",
-			wantStatus:      http.StatusOK,
-			wantContentType: "application/json",
-			wantBody: `{"users":[{"links":{"self":"/chronograf/v1/sources/1/users/strickland"},"name":"strickland","permissions":[{"scope":"all","allowed":["READ"]}]}]}
-`,
-		},
-		{
-			name: "All users for data source with roles",
-			args: args{
-				w: httptest.NewRecorder(),
-				r: httptest.NewRequest(
-					"GET",
-					"http://server.local/chronograf/v1/sources/1",
-					nil),
-			},
-			fields: fields{
-				UseAuth: true,
-				Logger:  log.New(log.DebugLevel),
-				SourcesStore: &mocks.SourcesStore{
-					GetF: func(ctx context.Context, ID int) (chronograf.Source, error) {
-						return chronograf.Source{
-							ID:       1,
-							Name:     "muh source",
-							Username: "name",
-							Password: "hunter2",
-							URL:      "http://localhost:8086",
-						}, nil
-					},
-				},
-				TimeSeries: &mocks.TimeSeries{
-					ConnectF: func(ctx context.Context, src *chronograf.Source) error {
-						return nil
-					},
-					RolesF: func(ctx context.Context) (chronograf.RolesStore, error) {
-						return nil, nil
-					},
-					UsersF: func(ctx context.Context) chronograf.UsersStore {
-						return &mocks.UsersStore{
-							AllF: func(ctx context.Context) ([]chronograf.User, error) {
-								return []chronograf.User{
-									{
-										Name:   "strickland",
-										Passwd: "discipline",
-										Permissions: chronograf.Permissions{
-											{
-												Scope:   chronograf.AllScope,
-												Allowed: chronograf.Allowances{"READ"},
-											},
-										},
-									},
-								}, nil
-							},
-						}
-					},
-				},
-			},
-			ID:              "1",
-			wantStatus:      http.StatusOK,
-			wantContentType: "application/json",
-			wantBody: `{"users":[{"links":{"self":"/chronograf/v1/sources/1/users/strickland"},"name":"strickland","permissions":[{"scope":"all","allowed":["READ"]}],"roles":[]}]}
-`,
-		},
-	}
-	for _, tt := range tests {
-		tt.args.r = tt.args.r.WithContext(httprouter.WithParams(
-			context.Background(),
-			httprouter.Params{
-				{
-					Key:   "id",
-					Value: tt.ID,
-				},
-			}))
-		h := &server.Service{
-			SourcesStore:     tt.fields.SourcesStore,
-			TimeSeriesClient: tt.fields.TimeSeries,
-			Logger:           tt.fields.Logger,
-			UseAuth:          tt.fields.UseAuth,
-		}
-
-		h.SourceUsers(tt.args.w, tt.args.r)
-		resp := tt.args.w.Result()
-		content := resp.Header.Get("Content-Type")
-		body, _ := ioutil.ReadAll(resp.Body)
-
-		if resp.StatusCode != tt.wantStatus {
-			t.Errorf("%q. SourceUsers() = %v, want %v", tt.name, resp.StatusCode, tt.wantStatus)
-		}
-		if tt.wantContentType != "" && content != tt.wantContentType {
-			t.Errorf("%q. SourceUsers() = %v, want %v", tt.name, content, tt.wantContentType)
-		}
-		if tt.wantBody != "" && string(body) != tt.wantBody {
-			t.Errorf("%q. SourceUsers() = \n***%v***\n,\nwant\n***%v***", tt.name, string(body), tt.wantBody)
-		}
-	}
-}
-
-func TestService_SourceUserID(t *testing.T) {
-	type fields struct {
-		SourcesStore chronograf.SourcesStore
-		TimeSeries   server.TimeSeriesClient
-		Logger       chronograf.Logger
-		UseAuth      bool
-	}
-	type args struct {
-		w *httptest.ResponseRecorder
-		r *http.Request
-	}
-	tests := []struct {
-		name            string
-		fields          fields
-		args            args
-		ID              string
-		UID             string
-		wantStatus      int
-		wantContentType string
-		wantBody        string
-	}{
-		{
-			name: "Single user for data source",
-			args: args{
-				w: httptest.NewRecorder(),
-				r: httptest.NewRequest(
-					"GET",
-					"http://server.local/chronograf/v1/sources/1",
-					nil),
-			},
-			fields: fields{
-				UseAuth: true,
-				Logger:  log.New(log.DebugLevel),
-				SourcesStore: &mocks.SourcesStore{
-					GetF: func(ctx context.Context, ID int) (chronograf.Source, error) {
-						return chronograf.Source{
-							ID:       1,
-							Name:     "muh source",
-							Username: "name",
-							Password: "hunter2",
-							URL:      "http://localhost:8086",
-						}, nil
-					},
-				},
-				TimeSeries: &mocks.TimeSeries{
-					ConnectF: func(ctx context.Context, src *chronograf.Source) error {
-						return nil
-					},
-					RolesF: func(ctx context.Context) (chronograf.RolesStore, error) {
-						return nil, fmt.Errorf("no roles")
-					},
-					UsersF: func(ctx context.Context) chronograf.UsersStore {
-						return &mocks.UsersStore{
-							GetF: func(ctx context.Context, uid string) (*chronograf.User, error) {
-								return &chronograf.User{
-									Name:   "strickland",
-									Passwd: "discipline",
-									Permissions: chronograf.Permissions{
-										{
-											Scope:   chronograf.AllScope,
-											Allowed: chronograf.Allowances{"READ"},
-										},
-									},
-								}, nil
-							},
-						}
-					},
-				},
-			},
-			ID:              "1",
-			UID:             "strickland",
-			wantStatus:      http.StatusOK,
-			wantContentType: "application/json",
-			wantBody: `{"links":{"self":"/chronograf/v1/sources/1/users/strickland"},"name":"strickland","permissions":[{"scope":"all","allowed":["READ"]}]}
-`,
-		},
-		{
-			name: "Single user for data source",
-			args: args{
-				w: httptest.NewRecorder(),
-				r: httptest.NewRequest(
-					"GET",
-					"http://server.local/chronograf/v1/sources/1",
-					nil),
-			},
-			fields: fields{
-				UseAuth: true,
-				Logger:  log.New(log.DebugLevel),
-				SourcesStore: &mocks.SourcesStore{
-					GetF: func(ctx context.Context, ID int) (chronograf.Source, error) {
-						return chronograf.Source{
-							ID:       1,
-							Name:     "muh source",
-							Username: "name",
-							Password: "hunter2",
-							URL:      "http://localhost:8086",
-						}, nil
-					},
-				},
-				TimeSeries: &mocks.TimeSeries{
-					ConnectF: func(ctx context.Context, src *chronograf.Source) error {
-						return nil
-					},
-					RolesF: func(ctx context.Context) (chronograf.RolesStore, error) {
-						return nil, nil
-					},
-					UsersF: func(ctx context.Context) chronograf.UsersStore {
-						return &mocks.UsersStore{
-							GetF: func(ctx context.Context, uid string) (*chronograf.User, error) {
-								return &chronograf.User{
-									Name:   "strickland",
-									Passwd: "discipline",
-									Permissions: chronograf.Permissions{
-										{
-											Scope:   chronograf.AllScope,
-											Allowed: chronograf.Allowances{"READ"},
-										},
-									},
-								}, nil
-							},
-						}
-					},
-				},
-			},
-			ID:              "1",
-			UID:             "strickland",
-			wantStatus:      http.StatusOK,
-			wantContentType: "application/json",
-			wantBody: `{"links":{"self":"/chronograf/v1/sources/1/users/strickland"},"name":"strickland","permissions":[{"scope":"all","allowed":["READ"]}],"roles":[]}
-`,
-		},
-	}
-	for _, tt := range tests {
-		tt.args.r = tt.args.r.WithContext(httprouter.WithParams(
-			context.Background(),
-			httprouter.Params{
-				{
-					Key:   "id",
-					Value: tt.ID,
-				},
-			}))
-		h := &server.Service{
-			SourcesStore:     tt.fields.SourcesStore,
-			TimeSeriesClient: tt.fields.TimeSeries,
-			Logger:           tt.fields.Logger,
-			UseAuth:          tt.fields.UseAuth,
-		}
-
-		h.SourceUserID(tt.args.w, tt.args.r)
-		resp := tt.args.w.Result()
-		content := resp.Header.Get("Content-Type")
-		body, _ := ioutil.ReadAll(resp.Body)
-
-		if resp.StatusCode != tt.wantStatus {
-			t.Errorf("%q. SourceUserID() = %v, want %v", tt.name, resp.StatusCode, tt.wantStatus)
-		}
-		if tt.wantContentType != "" && content != tt.wantContentType {
-			t.Errorf("%q. SourceUserID() = %v, want %v", tt.name, content, tt.wantContentType)
-		}
-		if tt.wantBody != "" && string(body) != tt.wantBody {
-			t.Errorf("%q. SourceUserID() = \n***%v***\n,\nwant\n***%v***", tt.name, string(body), tt.wantBody)
-		}
-	}
-}
-
-func TestService_RemoveSourceUser(t *testing.T) {
-	type fields struct {
-		SourcesStore chronograf.SourcesStore
-		TimeSeries   server.TimeSeriesClient
-		Logger       chronograf.Logger
-		UseAuth      bool
-	}
-	type args struct {
-		w *httptest.ResponseRecorder
-		r *http.Request
-	}
-	tests := []struct {
-		name            string
-		fields          fields
-		args            args
-		ID              string
-		UID             string
-		wantStatus      int
-		wantContentType string
-		wantBody        string
-	}{
-		{
-			name: "Delete user for data source",
-			args: args{
-				w: httptest.NewRecorder(),
-				r: httptest.NewRequest(
-					"GET",
-					"http://server.local/chronograf/v1/sources/1",
-					nil),
-			},
-			fields: fields{
-				UseAuth: true,
-				Logger:  log.New(log.DebugLevel),
-				SourcesStore: &mocks.SourcesStore{
-					GetF: func(ctx context.Context, ID int) (chronograf.Source, error) {
-						return chronograf.Source{
-							ID:       1,
-							Name:     "muh source",
-							Username: "name",
-							Password: "hunter2",
-							URL:      "http://localhost:8086",
-						}, nil
-					},
-				},
-				TimeSeries: &mocks.TimeSeries{
-					ConnectF: func(ctx context.Context, src *chronograf.Source) error {
-						return nil
-					},
-					UsersF: func(ctx context.Context) chronograf.UsersStore {
-						return &mocks.UsersStore{
-							DeleteF: func(ctx context.Context, u *chronograf.User) error {
-								return nil
-							},
-						}
-					},
-				},
-			},
-			ID:         "1",
-			UID:        "strickland",
+			id:         "1339",
 			wantStatus: http.StatusNoContent,
 		},
 	}
 	for _, tt := range tests {
-		tt.args.r = tt.args.r.WithContext(httprouter.WithParams(
-			context.Background(),
-			httprouter.Params{
-				{
-					Key:   "id",
-					Value: tt.ID,
-				},
-			}))
-		h := &server.Service{
-			SourcesStore:     tt.fields.SourcesStore,
-			TimeSeriesClient: tt.fields.TimeSeries,
-			Logger:           tt.fields.Logger,
-			UseAuth:          tt.fields.UseAuth,
-		}
-		h.RemoveSourceUser(tt.args.w, tt.args.r)
-		resp := tt.args.w.Result()
-		content := resp.Header.Get("Content-Type")
-		body, _ := ioutil.ReadAll(resp.Body)
+		t.Run(tt.name, func(t *testing.T) {
+			s := &Service{
+				UsersStore: tt.fields.UsersStore,
+				Logger:     tt.fields.Logger,
+			}
 
-		if resp.StatusCode != tt.wantStatus {
-			t.Errorf("%q. RemoveSourceUser() = %v, want %v", tt.name, resp.StatusCode, tt.wantStatus)
-		}
-		if tt.wantContentType != "" && content != tt.wantContentType {
-			t.Errorf("%q. RemoveSourceUser() = %v, want %v", tt.name, content, tt.wantContentType)
-		}
-		if tt.wantBody != "" && string(body) != tt.wantBody {
-			t.Errorf("%q. RemoveSourceUser() = \n***%v***\n,\nwant\n***%v***", tt.name, string(body), tt.wantBody)
-		}
+			tt.args.r = tt.args.r.WithContext(httprouter.WithParams(
+				context.Background(),
+				httprouter.Params{
+					{
+						Key:   "id",
+						Value: tt.id,
+					},
+				},
+			))
+
+			s.RemoveUser(tt.args.w, tt.args.r)
+
+			resp := tt.args.w.Result()
+
+			if resp.StatusCode != tt.wantStatus {
+				t.Errorf("%q. RemoveUser() = %v, want %v", tt.name, resp.StatusCode, tt.wantStatus)
+			}
+		})
 	}
 }
 
-func TestService_UpdateSourceUser(t *testing.T) {
+func TestService_UpdateUser(t *testing.T) {
 	type fields struct {
-		SourcesStore chronograf.SourcesStore
-		TimeSeries   server.TimeSeriesClient
-		Logger       chronograf.Logger
-		UseAuth      bool
+		UsersStore chronograf.UsersStore
+		Logger     chronograf.Logger
 	}
 	type args struct {
-		w *httptest.ResponseRecorder
-		r *http.Request
+		w    *httptest.ResponseRecorder
+		r    *http.Request
+		user *chronograf.User
 	}
 	tests := []struct {
 		name            string
 		fields          fields
 		args            args
-		ID              string
-		UID             string
+		id              string
 		wantStatus      int
 		wantContentType string
 		wantBody        string
 	}{
 		{
-			name: "Update user password for data source",
-			args: args{
-				w: httptest.NewRecorder(),
-				r: httptest.NewRequest(
-					"POST",
-					"http://server.local/chronograf/v1/sources/1",
-					ioutil.NopCloser(
-						bytes.NewReader([]byte(`{"name": "marty", "password": "the_lake"}`)))),
-			},
+			name: "Update a Chronograf user",
 			fields: fields{
-				UseAuth: true,
-				Logger:  log.New(log.DebugLevel),
-				SourcesStore: &mocks.SourcesStore{
-					GetF: func(ctx context.Context, ID int) (chronograf.Source, error) {
-						return chronograf.Source{
-							ID:       1,
-							Name:     "muh source",
-							Username: "name",
-							Password: "hunter2",
-							URL:      "http://localhost:8086",
-						}, nil
-					},
-				},
-				TimeSeries: &mocks.TimeSeries{
-					ConnectF: func(ctx context.Context, src *chronograf.Source) error {
+				Logger: log.New(log.DebugLevel),
+				UsersStore: &mocks.UsersStore{
+					UpdateF: func(ctx context.Context, user *chronograf.User) error {
 						return nil
 					},
-					RolesF: func(ctx context.Context) (chronograf.RolesStore, error) {
-						return nil, fmt.Errorf("no roles")
-					},
-					UsersF: func(ctx context.Context) chronograf.UsersStore {
-						return &mocks.UsersStore{
-							UpdateF: func(ctx context.Context, u *chronograf.User) error {
-								return nil
-							},
-							GetF: func(ctx context.Context, name string) (*chronograf.User, error) {
-								return &chronograf.User{
-									Name: "marty",
-								}, nil
-							},
+					GetF: func(ctx context.Context, ID string) (*chronograf.User, error) {
+						switch ID {
+						case "1336":
+							return &chronograf.User{
+								ID:       1336,
+								Name:     "bobbetta2",
+								Provider: "GitHub",
+								Scheme:   "OAuth2",
+							}, nil
+						default:
+							return nil, fmt.Errorf("User with ID %v not found", ID)
 						}
 					},
 				},
 			},
-			ID:              "1",
-			UID:             "marty",
-			wantStatus:      http.StatusOK,
-			wantContentType: "application/json",
-			wantBody: `{"links":{"self":"/chronograf/v1/sources/1/users/marty"},"name":"marty","permissions":[]}
-`,
-		},
-		{
-			name: "Update user password for data source with roles",
 			args: args{
 				w: httptest.NewRecorder(),
 				r: httptest.NewRequest(
-					"POST",
-					"http://server.local/chronograf/v1/sources/1",
-					ioutil.NopCloser(
-						bytes.NewReader([]byte(`{"name": "marty", "password": "the_lake"}`)))),
-			},
-			fields: fields{
-				UseAuth: true,
-				Logger:  log.New(log.DebugLevel),
-				SourcesStore: &mocks.SourcesStore{
-					GetF: func(ctx context.Context, ID int) (chronograf.Source, error) {
-						return chronograf.Source{
-							ID:       1,
-							Name:     "muh source",
-							Username: "name",
-							Password: "hunter2",
-							URL:      "http://localhost:8086",
-						}, nil
-					},
-				},
-				TimeSeries: &mocks.TimeSeries{
-					ConnectF: func(ctx context.Context, src *chronograf.Source) error {
-						return nil
-					},
-					RolesF: func(ctx context.Context) (chronograf.RolesStore, error) {
-						return nil, nil
-					},
-					UsersF: func(ctx context.Context) chronograf.UsersStore {
-						return &mocks.UsersStore{
-							UpdateF: func(ctx context.Context, u *chronograf.User) error {
-								return nil
-							},
-							GetF: func(ctx context.Context, name string) (*chronograf.User, error) {
-								return &chronograf.User{
-									Name: "marty",
-								}, nil
-							},
-						}
-					},
+					"PATCH",
+					"http://any.url",
+					nil,
+				),
+				user: &chronograf.User{
+					ID:       1336,
+					Name:     "bobbetta2",
+					Provider: "GitHub",
+					Scheme:   "OAuth2",
 				},
 			},
-			ID:              "1",
-			UID:             "marty",
+			id:              "1336",
 			wantStatus:      http.StatusOK,
 			wantContentType: "application/json",
-			wantBody: `{"links":{"self":"/chronograf/v1/sources/1/users/marty"},"name":"marty","permissions":[],"roles":[]}
-`,
-		},
-		{
-			name: "Invalid update JSON",
-			args: args{
-				w: httptest.NewRecorder(),
-				r: httptest.NewRequest(
-					"POST",
-					"http://server.local/chronograf/v1/sources/1",
-					ioutil.NopCloser(
-						bytes.NewReader([]byte(`{"name": "marty"}`)))),
-			},
-			fields: fields{
-				UseAuth: true,
-				Logger:  log.New(log.DebugLevel),
-			},
-			ID:              "1",
-			UID:             "marty",
-			wantStatus:      http.StatusUnprocessableEntity,
-			wantContentType: "application/json",
-			wantBody:        `{"code":422,"message":"No fields to update"}`,
+			wantBody:        `{"id":1336,"name":"bobbetta2","provider":"GitHub","scheme":"OAuth2","links":{"self":"/chronograf/v1/users/1336"}}`,
 		},
 	}
 	for _, tt := range tests {
-		tt.args.r = tt.args.r.WithContext(httprouter.WithParams(
-			context.Background(),
-			httprouter.Params{
-				{
-					Key:   "id",
-					Value: tt.ID,
-				},
-				{
-					Key:   "uid",
-					Value: tt.UID,
-				},
-			}))
-		h := &server.Service{
-			SourcesStore:     tt.fields.SourcesStore,
-			TimeSeriesClient: tt.fields.TimeSeries,
-			Logger:           tt.fields.Logger,
-			UseAuth:          tt.fields.UseAuth,
-		}
-		h.UpdateSourceUser(tt.args.w, tt.args.r)
-		resp := tt.args.w.Result()
-		content := resp.Header.Get("Content-Type")
-		body, _ := ioutil.ReadAll(resp.Body)
+		t.Run(tt.name, func(t *testing.T) {
+			s := &Service{
+				UsersStore: tt.fields.UsersStore,
+				Logger:     tt.fields.Logger,
+			}
 
-		if resp.StatusCode != tt.wantStatus {
-			t.Errorf("%q. UpdateSourceUser() = %v, want %v", tt.name, resp.StatusCode, tt.wantStatus)
-		}
-		if tt.wantContentType != "" && content != tt.wantContentType {
-			t.Errorf("%q. UpdateSourceUser() = %v, want %v", tt.name, content, tt.wantContentType)
-		}
-		if tt.wantBody != "" && string(body) != tt.wantBody {
-			t.Errorf("%q. UpdateSourceUser() = \n***%v***\n,\nwant\n***%v***", tt.name, string(body), tt.wantBody)
-		}
+			tt.args.r = tt.args.r.WithContext(httprouter.WithParams(context.Background(),
+				httprouter.Params{
+					{
+						Key:   "id",
+						Value: tt.id,
+					},
+				}))
+			buf, _ := json.Marshal(tt.args.user)
+			tt.args.r.Body = ioutil.NopCloser(bytes.NewReader(buf))
+
+			s.UpdateUser(tt.args.w, tt.args.r)
+
+			resp := tt.args.w.Result()
+			content := resp.Header.Get("Content-Type")
+			body, _ := ioutil.ReadAll(resp.Body)
+
+			if resp.StatusCode != tt.wantStatus {
+				t.Errorf("%q. UpdateUser() = %v, want %v", tt.name, resp.StatusCode, tt.wantStatus)
+			}
+			if tt.wantContentType != "" && content != tt.wantContentType {
+				t.Errorf("%q. UpdateUser() = %v, want %v", tt.name, content, tt.wantContentType)
+			}
+			if eq, _ := jsonEqual(string(body), tt.wantBody); tt.wantBody != "" && !eq {
+				t.Errorf("%q. UpdateUser() = \n***%v***\n,\nwant\n***%v***", tt.name, string(body), tt.wantBody)
+			}
+		})
 	}
+}
+
+func jsonEqual(s1, s2 string) (eq bool, err error) {
+	var o1, o2 interface{}
+
+	if err = json.Unmarshal([]byte(s1), &o1); err != nil {
+		return
+	}
+	if err = json.Unmarshal([]byte(s2), &o2); err != nil {
+		return
+	}
+
+	return cmp.Equal(o1, o2), nil
 }
